Move referrals vm to module
