+++ settings.gradle.kts
@@ -58,6 +58,7 @@ include(":modules:features:search")
 include(":modules:features:settings")
 include(":modules:features:shared")
 include(":modules:features:reimagine")
+include(":modules:features:referrals")
 include(":modules:features:taskerplugin")
 include(":modules:features:widgets")
 include(":modules:features:kids")
@@ -74,7 +75,6 @@ include(":modules:services:media-noop")
 include(":modules:services:model")
 include(":modules:services:preferences")
 include(":modules:services:protobuf")
-include(":modules:features:referrals")
 include(":modules:services:repositories")
 include(":modules:services:sharing")
 include(":modules:services:servers")
